+++ lib/core/alice_core.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:alice/model/alice_http_error.dart';
 import 'package:alice/ui/alice_calls_list_screen.dart';
 import 'package:alice/model/alice_http_call.dart';
@@ -9,31 +11,36 @@ import 'package:rxdart/rxdart.dart';
 import 'package:shake/shake.dart';
 
 class AliceCore {
+  final bool showNotification;
+  final bool showInspectorOnShake;
+  final bool darkTheme;
+  final BehaviorSubject<List<AliceHttpCall>> callsSubject =
+      BehaviorSubject.seeded(List());
+  final String notificationIcon;
+
   FlutterLocalNotificationsPlugin _flutterLocalNotificationsPlugin;
   GlobalKey<NavigatorState> _navigatorKey;
-  bool _showNotification = false;
-  bool _showInspectorOnShake = false;
-  bool _isInspectorOpened = false;
   Brightness _brightness = Brightness.light;
+  bool _isInspectorOpened = false;
+  ShakeDetector _shakeDetector;
+  StreamSubscription _callsSubscription;
+  String _notificationMessage;
+  String _notificationMessageShown;
+  bool _notificationProcessing = false;
 
-  List<AliceHttpCall> calls;
-  PublishSubject<int> changesSubject;
-  PublishSubject<AliceHttpCall> callUpdateSubject;
-  ShakeDetector shakeDetector;
-
-  AliceCore(GlobalKey<NavigatorState> navigatorKey, bool showNotification,
-      bool showInspectorOnShake, bool darkTheme) {
-    _navigatorKey = navigatorKey;
-    calls = List();
-    changesSubject = PublishSubject();
-    callUpdateSubject = PublishSubject();
-    _showNotification = showNotification;
+  AliceCore(this._navigatorKey, this.showNotification,
+      this.showInspectorOnShake, this.darkTheme, this.notificationIcon)
+      : assert(showNotification != null, "showNotification can't be null"),
+        assert(
+            showInspectorOnShake != null, "showInspectorOnShake can't be null"),
+        assert(darkTheme != null, "darkTheme can't be null"),
+        assert(notificationIcon != null, "notificationIcon can't be null") {
     if (showNotification) {
       _initializeNotificationsPlugin();
+      _callsSubscription = callsSubject.listen((_) => _onCallsChanged());
     }
-    _showInspectorOnShake = showInspectorOnShake;
-    if (_showInspectorOnShake) {
-      shakeDetector = ShakeDetector.autoStart(
+    if (showInspectorOnShake) {
+      _shakeDetector = ShakeDetector.autoStart(
         onPhoneShake: () => navigateToCallListScreen(),
         shakeThresholdGravity: 5,
       );
@@ -41,24 +48,43 @@ class AliceCore {
     _brightness = darkTheme ? Brightness.dark : Brightness.light;
   }
 
-  dispose() {
-    changesSubject.close();
-    callUpdateSubject.close();
-    shakeDetector?.stopListening();
+  void dispose() {
+    callsSubject.close();
+    _shakeDetector?.stopListening();
+    _callsSubscription?.cancel();
   }
 
+  Brightness get brightness => _brightness;
+
   void _initializeNotificationsPlugin() {
-    _flutterLocalNotificationsPlugin = new FlutterLocalNotificationsPlugin();
+    _flutterLocalNotificationsPlugin = FlutterLocalNotificationsPlugin();
     var initializationSettingsAndroid =
-        new AndroidInitializationSettings("@mipmap/ic_launcher");
-    var initializationSettingsIOS = new IOSInitializationSettings();
-    var initializationSettings = new InitializationSettings(
+        new AndroidInitializationSettings(notificationIcon);
+    var initializationSettingsIOS = IOSInitializationSettings();
+    var initializationSettings = InitializationSettings(
         initializationSettingsAndroid, initializationSettingsIOS);
     _flutterLocalNotificationsPlugin.initialize(initializationSettings,
         onSelectNotification: _onSelectedNotification);
   }
 
+  void _onCallsChanged() async {
+    if (callsSubject.value.length > 0) {
+      _notificationMessage = _getNotificationMessage();
+      if (_notificationMessage != _notificationMessageShown &&
+          !_notificationProcessing) {
+        await _showLocalNotification();
+        _onCallsChanged();
+      }
+    }
+  }
+
+  void setNavigatorKey(GlobalKey<NavigatorState> navigatorKey) {
+    assert(navigatorKey != null, "navigatorKey can't be null");
+    this._navigatorKey = navigatorKey;
+  }
+
   Future _onSelectedNotification(String payload) {
+    assert(payload != null, "payload can't be null");
     navigateToCallListScreen();
     return Future.sync(() {});
   }
@@ -74,22 +100,64 @@ class AliceCore {
       _isInspectorOpened = true;
       Navigator.push(
         context,
-        MaterialPageRoute(builder: (context) => AliceCallsListScreen(this)),
+        MaterialPageRoute(
+          builder: (context) => AliceCallsListScreen(this),
+        ),
       ).then((onValue) => _isInspectorOpened = false);
     }
   }
 
-  BuildContext getContext() {
-    if (_navigatorKey != null &&
-        _navigatorKey.currentState != null &&
-        _navigatorKey.currentState.overlay != null) {
-      return _navigatorKey.currentState.overlay.context;
-    } else {
-      return null;
+  BuildContext getContext() => _navigatorKey?.currentState?.overlay?.context;
+
+  String _getNotificationMessage() {
+    List<AliceHttpCall> calls = callsSubject.value;
+    int successCalls = calls
+        .where((call) =>
+            call.response != null &&
+            call.response.status >= 200 &&
+            call.response.status < 300)
+        .toList()
+        .length;
+
+    int redirectCalls = calls
+        .where((call) =>
+            call.response != null &&
+            call.response.status >= 300 &&
+            call.response.status < 400)
+        .toList()
+        .length;
+
+    int errorCalls = calls
+        .where((call) =>
+            call.response != null &&
+            call.response.status >= 400 &&
+            call.response.status < 600)
+        .toList()
+        .length;
+
+    int loadingCalls = calls.where((call) => call.loading).toList().length;
+
+    StringBuffer notificationsMessage = StringBuffer();
+    if (loadingCalls > 0) {
+      notificationsMessage.write("Loading: $loadingCalls");
+      notificationsMessage.write(" | ");
     }
+    if (successCalls > 0) {
+      notificationsMessage.write("Success: $successCalls");
+      notificationsMessage.write(" | ");
+    }
+    if (redirectCalls > 0) {
+      notificationsMessage.write("Error: $redirectCalls");
+      notificationsMessage.write(" | ");
+    }
+    if (errorCalls > 0) {
+      notificationsMessage.write("Error: $errorCalls");
+    }
+    return notificationsMessage.toString();
   }
 
-  void _showLocalNotification() async {
+  Future _showLocalNotification() async {
+    _notificationProcessing = true;
     var channelId = "Alice";
     var channelName = "Alice";
     var channelDescription = "Alice";
@@ -101,20 +169,26 @@ class AliceCore {
     var iOSPlatformChannelSpecifics = new IOSNotificationDetails();
     var platformChannelSpecifics = new NotificationDetails(
         androidPlatformChannelSpecifics, iOSPlatformChannelSpecifics);
-
+    String message = _notificationMessage;
     await _flutterLocalNotificationsPlugin.show(
-        0, "Alice", "HTTP Requests: ${calls.length}", platformChannelSpecifics,
+        0,
+        "Alice (total: ${callsSubject.value.length} requests)",
+        message,
+        platformChannelSpecifics,
         payload: "");
+    _notificationMessageShown = message;
+    _notificationProcessing = false;
+    return;
   }
 
   void addCall(AliceHttpCall call) {
-    calls.add(call);
-    if (_showNotification) {
-      _showLocalNotification();
-    }
+    assert(call != null, "call can't be null");
+    callsSubject.add(callsSubject.value..add(call));
   }
 
   void addError(AliceHttpError error, int requestId) {
+    assert(error != null, "error can't be null");
+    assert(requestId != null, "requestId can't be null");
     AliceHttpCall selectedCall = _selectCall(requestId);
 
     if (selectedCall == null) {
@@ -123,11 +197,12 @@ class AliceCore {
     }
 
     selectedCall.error = error;
-    changesSubject.sink.add(requestId);
-    callUpdateSubject.sink.add(selectedCall);
+    callsSubject.add(List()..addAll(callsSubject.value));
   }
 
   void addResponse(AliceHttpResponse response, int requestId) {
+    assert(response != null, "response can't be null");
+    assert(requestId != null, "requestId can't be null");
     AliceHttpCall selectedCall = _selectCall(requestId);
 
     if (selectedCall == null) {
@@ -139,28 +214,18 @@ class AliceCore {
     selectedCall.duration = response.time.millisecondsSinceEpoch -
         selectedCall.request.time.millisecondsSinceEpoch;
 
-    changesSubject.sink.add(requestId);
-    callUpdateSubject.sink.add(selectedCall);
+    callsSubject.add(List()..addAll(callsSubject.value));
   }
 
   void removeCalls() {
-    calls = List();
-    changesSubject.sink.add(0);
+    callsSubject.add(List());
   }
 
-  AliceHttpCall _selectCall(int requestId) {
-    AliceHttpCall requestedCall;
-    calls.forEach((call) {
-      if (call.id == requestId) {
-        requestedCall = call;
-      }
-    });
-    return requestedCall;
-  }
+  AliceHttpCall _selectCall(int requestId) => callsSubject.value
+      .firstWhere((call) => call.id == requestId, orElse: null);
 
   void saveHttpRequests(BuildContext context) {
-    AliceSaveHelper.saveCalls(context, calls);
+    assert(context != null, "context can't be null");
+    AliceSaveHelper.saveCalls(context, callsSubject.value);
   }
-
-  Brightness get brightness => _brightness;
 }