+++ lib/core/alice_dio_interceptor.dart
@@ -1,6 +1,8 @@
 import 'dart:convert';
 
 import 'package:alice/core/alice_core.dart';
+import 'package:alice/model/alice_form_data_file.dart';
+import 'package:alice/model/alice_from_data_field.dart';
 import 'package:alice/model/alice_http_call.dart';
 import 'package:alice/model/alice_http_error.dart';
 import 'package:alice/model/alice_http_request.dart';
@@ -8,14 +10,14 @@ import 'package:alice/model/alice_http_response.dart';
 import 'package:dio/dio.dart';
 
 class AliceDioInterceptor extends InterceptorsWrapper {
-  AliceCore _aliceCore;
+  final AliceCore aliceCore;
 
-  AliceDioInterceptor(AliceCore aliceCore) {
-    _aliceCore = aliceCore;
-  }
+  AliceDioInterceptor(this.aliceCore)
+      : assert(aliceCore != null, "aliceCore can't be null");
 
   @override
-  onRequest(RequestOptions options) {
+  Future onRequest(RequestOptions options) {
+    assert(options != null, "options can't be null");
     AliceHttpCall call = new AliceHttpCall(options.hashCode);
 
     Uri uri = options.uri;
@@ -35,13 +37,36 @@ class AliceDioInterceptor extends InterceptorsWrapper {
 
     AliceHttpRequest request = AliceHttpRequest();
 
-    if (options.data == null) {
+    var data = options.data;
+    if (data == null) {
       request.size = 0;
       request.body = "";
     } else {
-      request.size = utf8.encode(options.data.toString()).length;
-      request.body = options.data;
+      if (data is FormData) {
+        request.body += "Form data";
+
+        if (data.fields?.isNotEmpty == true) {
+          List<AliceFormDataField> fields = List();
+          data.fields.forEach((entry) {
+            fields.add(AliceFormDataField(entry.key, entry.value));
+          });
+          request.formDataFields = fields;
+        }
+        if (data.files?.isNotEmpty == true) {
+          List<AliceFormDataFile> files = List();
+          data.files.forEach((entry) {
+            files.add(AliceFormDataFile(entry.value.filename,
+                entry.value.contentType.toString(), entry.value.length));
+          });
+
+          request.formDataFiles = files;
+        }
+      } else {
+        request.size = utf8.encode(data.toString()).length;
+        request.body = data;
+      }
     }
+
     request.time = DateTime.now();
     request.headers = options.headers;
     request.contentType = options.contentType.toString();
@@ -50,12 +75,13 @@ class AliceDioInterceptor extends InterceptorsWrapper {
     call.request = request;
     call.response = AliceHttpResponse();
 
-    _aliceCore.addCall(call);
+    aliceCore.addCall(call);
     return super.onRequest(options);
   }
 
   @override
-  onResponse(Response response) {
+  Future onResponse(Response response) {
+    assert(response != null, "response can't be null");
     var httpResponse = AliceHttpResponse();
     httpResponse.status = response.statusCode;
 
@@ -76,45 +102,46 @@ class AliceDioInterceptor extends InterceptorsWrapper {
     }
     httpResponse.headers = headers;
 
-    _aliceCore.addResponse(httpResponse, response.request.hashCode);
+    aliceCore.addResponse(httpResponse, response.request.hashCode);
     return super.onResponse(response);
   }
 
   @override
-  onError(DioError err) {
+  Future onError(DioError error) {
+    assert(error != null, "error can't be null");
     var httpError = AliceHttpError();
-    httpError.error = err.toString();
-    if (err is Error) {
-      var error = err as Error;
-      httpError.stackTrace = error.stackTrace;
+    httpError.error = error.toString();
+    if (error is Error) {
+      var basicError = error as Error;
+      httpError.stackTrace = basicError.stackTrace;
     }
 
-    _aliceCore.addError(httpError, err.request.hashCode);
+    aliceCore.addError(httpError, error.request.hashCode);
     var httpResponse = AliceHttpResponse();
     httpResponse.time = DateTime.now();
-    if (err.response == null) {
+    if (error.response == null) {
       httpResponse.status = -1;
-      _aliceCore.addResponse(httpResponse, err.request.hashCode);
+      aliceCore.addResponse(httpResponse, error.request.hashCode);
     } else {
-      httpResponse.status = err.response.statusCode;
+      httpResponse.status = error.response.statusCode;
 
-      if (err.response.data == null) {
+      if (error.response.data == null) {
         httpResponse.body = "";
         httpResponse.size = 0;
       } else {
-        httpResponse.body = err.response.data;
-        httpResponse.size = utf8.encode(err.response.data.toString()).length;
+        httpResponse.body = error.response.data;
+        httpResponse.size = utf8.encode(error.response.data.toString()).length;
       }
       Map<String, String> headers = Map();
-      if (err.response.headers != null) {
-        err.response.headers.forEach((header, values) {
+      if (error.response.headers != null) {
+        error.response.headers.forEach((header, values) {
           headers[header] = values.toString();
         });
       }
       httpResponse.headers = headers;
-      _aliceCore.addResponse(httpResponse, err.response.request.hashCode);
+      aliceCore.addResponse(httpResponse, error.response.request.hashCode);
     }
 
-    return super.onError(err);
+    return super.onError(error);
   }
 }