+++ lib/ui/alice_calls_list_screen.dart
@@ -18,6 +18,8 @@ class AliceCallsListScreen extends StatefulWidget {
 }
 
 class _AliceCallsListScreenState extends State<AliceCallsListScreen> {
+  AliceCore get aliceCore => widget._aliceCore;
+
   List<AliceMenuItem> _menuItems = List();
 
   _AliceCallsListScreenState() {
@@ -29,33 +31,38 @@ class _AliceCallsListScreenState extends State<AliceCallsListScreen> {
   @override
   Widget build(BuildContext context) {
     return Theme(
-        data: ThemeData(brightness: widget._aliceCore.brightness),
-        child: Scaffold(
-            appBar: AppBar(
-              title: Text("Alice - HTTP Inspector - Calls"),
-              actions: [
-                PopupMenuButton<AliceMenuItem>(
-                  onSelected: (AliceMenuItem item) => _onMenuItemSelected(item),
-                  itemBuilder: (BuildContext context) {
-                    return _menuItems.map((AliceMenuItem item) {
-                      return PopupMenuItem<AliceMenuItem>(
-                          value: item,
-                          child: Row(children: [
-                            Icon(
-                              item.iconData,
-                              color: Colors.lightBlue,
-                            ),
-                            Padding(
-                              padding: EdgeInsets.only(left: 10),
-                            ),
-                            Text(item.title)
-                          ]));
-                    }).toList();
-                  },
-                ),
-              ],
-            ),
-            body: _getCallsList()));
+      data: ThemeData(brightness: widget._aliceCore.brightness),
+      child: Scaffold(
+        appBar: AppBar(
+          title: Text("Alice - HTTP Inspector - Calls"),
+          actions: [_buildMenuButton()],
+        ),
+        body: _buildCallsListWrapper(),
+      ),
+    );
+  }
+
+  Widget _buildMenuButton() {
+    return PopupMenuButton<AliceMenuItem>(
+      onSelected: (AliceMenuItem item) => _onMenuItemSelected(item),
+      itemBuilder: (BuildContext context) {
+        return _menuItems.map((AliceMenuItem item) {
+          return PopupMenuItem<AliceMenuItem>(
+            value: item,
+            child: Row(children: [
+              Icon(
+                item.iconData,
+                color: Colors.lightBlue,
+              ),
+              Padding(
+                padding: EdgeInsets.only(left: 10),
+              ),
+              Text(item.title)
+            ]),
+          );
+        }).toList();
+      },
+    );
   }
 
   void _onMenuItemSelected(AliceMenuItem menuItem) {
@@ -70,83 +77,87 @@ class _AliceCallsListScreenState extends State<AliceCallsListScreen> {
     }
   }
 
-  Widget _getCallsList() {
-    return StreamBuilder(
-        stream: widget._aliceCore.changesSubject.stream,
-        builder: (context, AsyncSnapshot<int> snapshot) {
-          if (widget._aliceCore.calls.length == 0) {
-            return Container(
-                margin: EdgeInsets.all(5),
-                child: Center(
-                  child: Column(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: [
-                        Icon(
-                          Icons.error_outline,
-                          color: Colors.orange,
-                        ),
-                        Padding(
-                          padding: EdgeInsets.only(top: 5),
-                        ),
-                        Text(
-                          "There are no calls to show",
-                          style: TextStyle(fontSize: 18),
-                        ),
-                        Padding(
-                          padding: EdgeInsets.only(top: 5),
-                        ),
-                        Text(
-                          "You have not send any http call or your Alice configuration is invalid.",
-                          style: TextStyle(fontSize: 12),
-                        )
-                      ]),
-                ));
-          } else {
-            return ListView(
-              children: _getListElements(),
-            );
-          }
-        });
+  Widget _buildCallsListWrapper() {
+    return StreamBuilder<List<AliceHttpCall>>(
+      stream: aliceCore.callsSubject,
+      builder: (context, snapshot) {
+        if (snapshot.hasData && snapshot.data.isNotEmpty) {
+          return _buildCallsListWidget(snapshot.data);
+        } else {
+          return _buildEmptyWidget();
+        }
+      },
+    );
   }
 
-  _getListElements() {
-    List<Widget> widgets = List();
-    widget._aliceCore.calls.forEach((AliceHttpCall call) {
-      widgets.add(AliceCallListItem(call, _onListItemClicked));
-    });
-    return widgets;
+  Widget _buildEmptyWidget() {
+    return Container(
+      margin: EdgeInsets.all(5),
+      child: Center(
+        child: Column(mainAxisAlignment: MainAxisAlignment.center, children: [
+          Icon(
+            Icons.error_outline,
+            color: Colors.orange,
+          ),
+          const SizedBox(height: 6),
+          Text(
+            "There are no calls to show",
+            style: TextStyle(fontSize: 18),
+          ),
+          const SizedBox(height: 6),
+          Text(
+            "You have not send any http call or your Alice configuration is invalid.",
+            style: TextStyle(fontSize: 12),
+          )
+        ]),
+      ),
+    );
+  }
+
+  Widget _buildCallsListWidget(List<AliceHttpCall> calls) {
+    return ListView.builder(
+      itemCount: calls.length,
+      itemBuilder: (context, index) {
+        return AliceCallListItem(calls[index], _onListItemClicked);
+      },
+    );
   }
 
   void _onListItemClicked(AliceHttpCall call) {
     Navigator.push(
       widget._aliceCore.getContext(),
       MaterialPageRoute(
-          builder: (context) =>
-              AliceCallDetailsScreen(call, widget._aliceCore)),
+        builder: (context) => AliceCallDetailsScreen(call, widget._aliceCore),
+      ),
     );
   }
 
   void _showRemoveDialog() {
     AliceAlertHelper.showAlert(
-        context, "Delete calls", "Do you want to delete http calls?",
-        firstButtonTitle: "No",
-        firstButtonAction: () => {},
-        secondButtonTitle: "Yes",
-        secondButtonAction: () => _removeCalls());
+      context,
+      "Delete calls",
+      "Do you want to delete http calls?",
+      firstButtonTitle: "No",
+      firstButtonAction: () => {},
+      secondButtonTitle: "Yes",
+      secondButtonAction: () => _removeCalls(),
+    );
   }
 
   void _removeCalls() {
-    widget._aliceCore.removeCalls();
+    aliceCore.removeCalls();
   }
 
   void _showStatsScreen() {
     Navigator.push(
-        widget._aliceCore.getContext(),
-        MaterialPageRoute(
-            builder: (context) => AliceStatsScreen(widget._aliceCore)));
+      aliceCore.getContext(),
+      MaterialPageRoute(
+        builder: (context) => AliceStatsScreen(widget._aliceCore),
+      ),
+    );
   }
 
   void _saveToFile() async {
-    widget._aliceCore.saveHttpRequests(context);
+    aliceCore.saveHttpRequests(context);
   }
 }